Remove fulfilment status and financial status from bold_checkout_order table.
