+++ Model/Order/PlaceOrder/CreateOrderFromPayload.php
@@ -124,14 +124,13 @@ class CreateOrderFromPayload
         $orderExtensionData = $this->orderExtensionDataFactory->create();
         $orderExtensionData->setPublicId($orderPayload->getPublicId());
         $orderExtensionData->setOrderId((int)$magentoOrder->getId());
-        $orderExtensionData->setFulfillmentStatus($orderPayload->getFulfillmentStatus());
-        $orderExtensionData->setFinancialStatus($orderPayload->getFinancialStatus());
-
         $this->eventManager->dispatch(
             'create_order_from_payload_extension_data_save_before',
-            ['orderPayload' => $orderPayload, 'orderExtensionData' => $orderExtensionData]
+            [
+                'orderPayload' => $orderPayload,
+                'orderExtensionData' => $orderExtensionData,
+            ]
         );
-
         $this->orderExtensionDataResource->save($orderExtensionData);
 
         return $magentoOrder;
