+++ etc/db_schema.xml
@@ -4,8 +4,6 @@
         <column xsi:type="int" name="id" unsigned="true" nullable="false" identity="true" comment="ID"/>
         <column xsi:type="int" name="order_id" unsigned="true" nullable="false" identity="false" comment="Magento Order ID"/>
         <column xsi:type="varchar" name="public_id" nullable="true" length="255" comment="Bold Order Public ID"/>
-        <column xsi:type="varchar" name="financial_status" nullable="true" length="255" comment="Bold Order Financial Status"/>
-        <column xsi:type="varchar" name="fulfillment_status" nullable="true" length="255" comment="Bold Order Fulfilment Status"/>
         <column xsi:type="smallint" name="is_delayed_capture" nullable="false" default="0" comment="Is Order Using Delayed Payment Capture."/>
         <constraint xsi:type="primary" referenceId="PRIMARY">
             <column name="id"/>
