+++ etc/di.xml
@@ -104,7 +104,6 @@
         <arguments>
             <argument name="code" xsi:type="const">Bold\Checkout\Model\Payment\Gateway\Service::CODE</argument>
             <argument name="valueHandlerPool" xsi:type="object">Bold\Checkout\Gateway\Config\ValueHandlerPool</argument>
-            <argument name="validatorPool" xsi:type="object">Bold\Checkout\Model\Payment\Gateway\Validator\ValidatorPool</argument>
             <argument name="commandPool" xsi:type="object">Bold\Checkout\Gateway\Command\CommandPool</argument>
             <argument name="formBlockType" xsi:type="object">Magento\Payment\Block\Form</argument>
             <argument name="infoBlockType" xsi:type="string">Bold\Checkout\Block\Info</argument>
@@ -129,15 +128,6 @@
     <virtualType name="Bold\Checkout\Model\Payment\Gateway\Config\Config" type="Magento\Payment\Gateway\Config\Config">
         <arguments>
             <argument name="methodCode" xsi:type="const">Bold\Checkout\Model\Payment\Gateway\Service::CODE</argument>
-        </arguments>
-    </virtualType>
-    <!-- Validator Pool-->
-    <virtualType name="Bold\Checkout\Model\Payment\Gateway\Validator\ValidatorPool" type="Magento\Payment\Gateway\Validator\ValidatorPool">
-        <arguments>
-            <argument name="validators" xsi:type="array">
-                <item name="public_order_id" xsi:type="string">Bold\Checkout\Model\Payment\Gateway\Validator\PublicOrderIdValidator</item>
-                <item name="module_enabled" xsi:type="string">Bold\Checkout\Model\Payment\Gateway\Validator\ModuleEnabledValidator</item>
-            </argument>
         </arguments>
     </virtualType>
     <!-- Command Pool -->
