+++ etc/extension_attributes.xml
@@ -1,22 +1,5 @@
 <?xml version="1.0"?>
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Api/etc/extension_attributes.xsd">
-    <extension_attributes for="Magento\Sales\Api\Data\OrderInterface">
-        <attribute code="public_id" type="string">
-            <join reference_table="bold_checkout_order" join_on_field="entity_id" reference_field="order_id">
-                <field>public_id</field>
-            </join>
-        </attribute>
-        <attribute code="financial_status" type="string">
-            <join reference_table="bold_checkout_order" join_on_field="entity_id" reference_field="order_id">
-                <field>financial_status</field>
-            </join>
-        </attribute>
-        <attribute code="fulfillment_status" type="string">
-            <join reference_table="bold_checkout_order" join_on_field="entity_id" reference_field="order_id">
-                <field>fulfillment_status</field>
-            </join>
-        </attribute>
-    </extension_attributes>
     <extension_attributes for="Magento\Sales\Api\Data\OrderItemInterface">
         <attribute code="product" type="Magento\Catalog\Api\Data\ProductInterface"/>
     </extension_attributes>
